Move ProtectedVisibility Mutators to dedicated file
