+++ tests/Fixtures/Autoloaded/ProtectedVisibility/pv-abstract-class-protected-method-m0.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ProtectedVisibilityAbstractClassProtectedMethod;
+
+abstract class Test
+{
+    private function foo(int $param, $test = 1) : bool
+    {
+        echo 1;
+        return false;
+    }
+}
