+++ tests/Fixtures/Autoloaded/ProtectedVisibility/pv-final-m0.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ProtectedVisibilityFinal;
+
+class Test
+{
+    private final function &foo(int $param, $test = 1) : bool
+    {
+        echo 1;
+        return false;
+    }
+}
