+++ tests/Fixtures/Autoloaded/ProtectedVisibility/pv-non-same-method-parent-m0.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace ProtectedNonSameAbstract;
+
+abstract class ProtectedNonSameAbstract
+{
+    protected abstract function foo();
+}
+class Child extends ProtectedNonSameAbstract
+{
+    protected function foo()
+    {
+    }
+    private function bar()
+    {
+    }
+}
