+++ tests/Fixtures/Autoloaded/ProtectedVisibility/pv-one-class-m0.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ProtectedVisibilityOneClass;
+
+class Test
+{
+    private function foo(int $param, $test = 1) : bool
+    {
+        echo 1;
+        return false;
+    }
+}
