+++ tests/Fixtures/Autoloaded/ProtectedVisibility/pv-same-method-grandparent-m0.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace ProtectedSameGrandParent;
+
+class SameGrandParent
+{
+    private function foo()
+    {
+    }
+}
+class SameParent extends SameGrandParent
+{
+}
+class Child extends SameParent
+{
+    protected function foo()
+    {
+    }
+}
