+++ tests/Fixtures/Autoloaded/ProtectedVisibility/pv-same-method-parent-m0.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace ProtectedSameParent;
+
+class SameParent
+{
+    private function foo()
+    {
+    }
+}
+class Child extends SameParent
+{
+    protected function foo()
+    {
+    }
+}
