+++ tests/Fixtures/Autoloaded/ProtectedVisibility/pv-static-m0.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ProtectedVisibilityStatic;
+
+class Test
+{
+    private static function foo(int $param, $test = 1) : bool
+    {
+        echo 1;
+        return false;
+    }
+}
